+++ Editor/Playa/Renderer/SerializedFieldRendererUIToolkit.cs
@@ -7,7 +7,6 @@ using SaintsField.Editor.Utils;
 using SaintsField.Playa;
 using UnityEditor;
 using UnityEditor.UIElements;
-using UnityEngine;
 using UnityEngine.UIElements;
 
 namespace SaintsField.Editor.Playa.Renderer